requirements.txt
streamlit
fastapi
psycopg2
uvicorn
bcrypt
python-jose
requests
python-dotenv
pydantic
pydantic-settings 